# swagger/app.yml
swagger: "2.0"
info:
  title: "My Flask App"
  description: "Documentation for my Flask app"
  version: "1.0"
basePath: "/api"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

paths:
  /login:
    post:
      tags:
        - "Authentication"
      summary: "Logs in a user with a username and password"
      consumes:
        - "application/json"
      parameters:
        - name: username
          in: body
          description: "Username of the user to log in"
          required: true
          schema:
            type: string
        - name: password
          in: body
          description: "Password of the user to log in"
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Success"
          schema:
            type: string
        400:
          description: "Invalid input"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"

  /signup:
    post:
      tags:
        - "Authentication"
      summary: "Registers a new user with a username and password"
      consumes:
        - "application/json"
      parameters:
        - name: username
          in: body
          description: "Username of the new user"
          required: true
          schema:
            type: string
        - name: password
          in: body
          description: "Password of the new user"
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Success"
          schema:
            type: string
        400:
          description: "Invalid input"
        500:
          description: "Internal server error"

  /users:
    get:
      tags:
        - "Users"
      summary: "Gets a list of all users"
      consumes:
        - "application/json"
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: pong
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"